+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `plasma.has`
+- `plasma.utils.deepEqual`
+- `plasma.utils.isFilledArray`
+
 ## [1.0.1] - 2016-02-17
 
 ### Fixed